requirements.txt
asgiref==3.8.1
Django
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
mccabe==0.7.0
packaging==24.1
pillow==11.0.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
rest-framework-simplejwt==0.0.2
setuptools==75.2.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
